Move the middleware file
